! This tinyint actually needs to be an int. Because this is a new-to-2.1 feature this should NOT be separately handled by the upgrader (i.e. the upgrader should NOT go out of its way to fix this). You may need to update your test installations by hand.

+++ other/upgrade_2-1_postgresql.sql
@@ -463,7 +463,7 @@ CREATE TABLE {$db_prefix}user_alerts (
   content_type varchar(255) NOT NULL default '',
   content_id int NOT NULL default '0',
   content_action varchar(255) NOT NULL default '',
-  is_read smallint NOT NULL default '0',
+  is_read int NOT NULL default '0',
   extra text NOT NULL,
   PRIMARY KEY (id_alert)
 );
